requirements.txt
opencv-python
numpy
matplotlib
torch
torchvision
opencv-python-headless
pillow
tqdm
